chore(indexer): security improvements (#323)
